+++ knytt/objects/bank15/FadeBlock.cs
@@ -7,7 +7,7 @@ public partial class FadeBlock : GDKnyttBaseObject
     public override void _PhysicsProcess(double delta)
     {
         base._PhysicsProcess(delta);
-        pos += delta * (random.Next(5) - 1) * 50 * Mathf.Pi / 90;
+        pos += (float)delta * (random.Next(5) - 1) * 50 * Mathf.Pi / 90;
         Modulate = new Color(1, 1, 1, 0.16f + 0.16f * Mathf.Sin(pos));
     }
 }
