+++ knytt/objects/bullets/SmallGlowingBullet.cs
@@ -4,7 +4,7 @@ public partial class SmallGlowingBullet : BaseBullet
     public override void _PhysicsProcess(double delta)
     {
         if (!Enabled) { return; }
-        VelocityMMF2 += delta * 20;
+        VelocityMMF2 += (float)delta * 20;
         base._PhysicsProcess(delta);
     }
 }
