+++ knytt/objects/bullets/SuperBullet.cs
@@ -5,7 +5,7 @@ public partial class SuperBullet : BaseBullet
     public override void _PhysicsProcess(double delta)
     {
         base._PhysicsProcess(delta);
-        Translate(new Vector2(GDKnyttDataStore.random.Next(3) - 1, GDKnyttDataStore.random.Next(3) - 1) * delta * 30);
+        Translate(new Vector2(GDKnyttDataStore.random.Next(3) - 1, GDKnyttDataStore.random.Next(3) - 1) * (float)delta * 30);
     }
 
     private void onDisappear(int limit)
